feat: transport task.registration etc to study.registration & remove observation.registration
