--- observations/fixtures/registrations.json
@@ -1,67 +1,0 @@
-[
-{
-    "model": "observations.registration",
-    "pk": 1,
-    "fields": {
-        "order": 1,
-        "description": "logdata",
-        "description_nl": "logdata",
-        "description_en": "log data",
-        "needs_details": false,
-        "requires_review": false,
-        "is_recording": false
-    }
-},
-{
-    "model": "observations.registration",
-    "pk": 2,
-    "fields": {
-        "order": 2,
-        "description": "audio-opname",
-        "description_nl": "audio-opname",
-        "description_en": "audio recording",
-        "needs_details": false,
-        "requires_review": false,
-        "is_recording": true
-    }
-},
-{
-    "model": "observations.registration",
-    "pk": 3,
-    "fields": {
-        "order": 3,
-        "description": "video-opname",
-        "description_nl": "video-opname",
-        "description_en": "video recording",
-        "needs_details": false,
-        "requires_review": true,
-        "is_recording": true
-    }
-},
-{
-    "model": "observations.registration",
-    "pk": 4,
-    "fields": {
-        "order": 4,
-        "description": "notities",
-        "description_nl": "notities",
-        "description_en": "notes",
-        "needs_details": false,
-        "requires_review": false,
-        "is_recording": false
-    }
-},
-{
-    "model": "observations.registration",
-    "pk": 5,
-    "fields": {
-        "order": 5,
-        "description": "anders",
-        "description_nl": "anders",
-        "description_en": "otherwise",
-        "needs_details": true,
-        "requires_review": false,
-        "is_recording": false
-    }
-}
-]
