--- observations/translation.py
@@ -1,8 +1,0 @@
-from modeltranslation.translator import register, TranslationOptions
-
-from .models import Registration
-
-
-@register(Registration)
-class RegistrationTranslationOptions(TranslationOptions):
-    fields = ("description",)
