+++ studies/fixtures/registrations.json
@@ -1,6 +1,6 @@
 [
 {
-    "model": "tasks.registration",
+    "model": "studies.registration",
     "pk": 1,
     "fields": {
         "order": 1,
@@ -16,7 +16,7 @@
     }
 },
 {
-    "model": "tasks.registration",
+    "model": "studies.registration",
     "pk": 2,
     "fields": {
         "order": 2,
@@ -32,7 +32,7 @@
     }
 },
 {
-    "model": "tasks.registration",
+    "model": "studies.registration",
     "pk": 3,
     "fields": {
         "order": 3,
@@ -48,7 +48,7 @@
     }
 },
 {
-    "model": "tasks.registration",
+    "model": "studies.registration",
     "pk": 4,
     "fields": {
         "order": 4,
@@ -64,7 +64,7 @@
     }
 },
 {
-    "model": "tasks.registration",
+    "model": "studies.registration",
     "pk": 5,
     "fields": {
         "order": 5,
@@ -80,7 +80,7 @@
     }
 },
 {
-    "model": "tasks.registration",
+    "model": "studies.registration",
     "pk": 6,
     "fields": {
         "order": 6,
@@ -96,7 +96,7 @@
     }
 },
 {
-    "model": "tasks.registration",
+    "model": "studies.registration",
     "pk": 7,
     "fields": {
         "order": 7,
@@ -112,7 +112,7 @@
     }
 },
 {
-    "model": "tasks.registration",
+    "model": "studies.registration",
     "pk": 8,
     "fields": {
         "order": 8,
@@ -128,7 +128,7 @@
     }
 },
 {
-    "model": "tasks.registration",
+    "model": "studies.registration",
     "pk": 9,
     "fields": {
         "order": 9,
