+++ tasks/forms.py
@@ -91,10 +91,10 @@ class TaskForm(SoftValidationMixin, ConditionalModelForm):
             "description",
             "repeats",
             "duration",
-            "registrations",
-            "registrations_details",
-            "registration_kinds",
-            "registration_kinds_details",
+            # "registrations",
+            # "registrations_details",
+            # "registration_kinds",
+            # "registration_kinds_details",
             "feedback",
             "feedback_details",
         ]
@@ -110,8 +110,8 @@ geef dan <strong>het redelijkerwijs te verwachten maximum op</strong>."
             ),
         }
         widgets = {
-            "registrations": BootstrapCheckboxSelectMultiple(),
-            "registration_kinds": BootstrapCheckboxSelectMultiple(),
+            # "registrations": BootstrapCheckboxSelectMultiple(),
+            # "registration_kinds": BootstrapCheckboxSelectMultiple(),
             "feedback": BootstrapRadioSelect(choices=YES_NO),
         }
 
@@ -139,18 +139,18 @@ geef dan <strong>het redelijkerwijs te verwachten maximum op</strong>."
         )
 
         self.check_empty(cleaned_data, "feedback")
-        self.check_dependency_multiple(
-            cleaned_data, "registrations", "needs_kind", "registration_kinds"
-        )
-        self.check_dependency_multiple(
-            cleaned_data, "registrations", "needs_details", "registrations_details"
-        )
-        self.check_dependency_multiple(
-            cleaned_data,
-            "registration_kinds",
-            "needs_details",
-            "registration_kinds_details",
-        )
+        # self.check_dependency_multiple(
+        #     cleaned_data, "registrations", "needs_kind", "registration_kinds"
+        # )
+        # self.check_dependency_multiple(
+        #     cleaned_data, "registrations", "needs_details", "registrations_details"
+        # )
+        # self.check_dependency_multiple(
+        #     cleaned_data,
+        #     "registration_kinds",
+        #     "needs_details",
+        #     "registration_kinds_details",
+        # )
         self.check_dependency(cleaned_data, "feedback", "feedback_details")
